Version 1.1.3 across the board
